Added methods to read the composer.json file and query the WordPress.org theme and plugin API
Ping the WordPress.org plugin and theme API to get information about installed plugins. If a plugin us available in the WordPress.org repo, that plugin or theme can be added to the composer.json file.

WP-CLI commands available for the plugin including the ability to add the installed plugins and themes to the composer.json file.

Todo:
- Write plugins and themes to composer.json file

- Create an admin panel where users can manage settings for the plugin