Software can now work as a WP-CLI package and a WordPress plugin
Fixed errors when calling certain commands of the plugin that would cause WP-CLI to throw a fatal error. This was caused by certain WordPress functions not being defined, which would cause the command `wp plugin path` command not to work. This plugin makes use of the that command to set the installer-paths in the composer.json file.