Fixed bugs that caused composer.json file to be indented with tabs and fixed error with the install command
If a site already had a composer.json file and it was indented with spaces, the file would be indented with tabs once the plugin and theme dependencies were added to the composer.json file.

Fixed error that caused the "wp composer install plugins" and the "wp composer install themes" not to work.